Static analysis tools suggestions

+++ src/LevelSearchLayer.cpp
@@ -23,10 +23,10 @@ class $modify(GDDLLevelSearchLayer, LevelSearchLayer) {
         button->setID("gddl_search_button"_spr);
         getChildByIDRecursive("other-filter-menu")->addChild(button);
         auto buttonAbove = getChildByIDRecursive("lists-button");
-        button->setPosition({buttonAbove->getPositionX(), buttonAbove->getPositionY()-(50.0f * (Mod::get()->getSettingValue<int64_t>("move-gddl-search-button-down") + 1))});
+        button->setPosition({buttonAbove->getPositionX(), buttonAbove->getPositionY()-(50.0f * (static_cast<float>(Mod::get()->getSettingValue<int64_t>("move-gddl-search-button-down") + 1)))});
     }
 
-    void onGDDLSearch(CCObject* sender) {
+    void onGDDLSearch(CCObject* sender) { // NOLINT(*-convert-member-functions-to-static)
         GDDLSearchLayer::create()->show();
     }
 };
